Fix false lock detector for GPS 20ms, add false lock detector for GLO 10ms

+++ src/track.h
@@ -827,7 +827,7 @@ u8 tp_get_current_cycle_duration(tp_tm_e tracking_mode, u8 cycle_no);
 u32 tp_get_rollover_cycle_duration(tp_tm_e tracking_mode, u8 cycle_no);
 u8 tp_get_cn0_ms(tp_tm_e tracking_mode);
 u8 tp_get_ld_ms(tp_tm_e tracking_mode);
-u8 tp_get_alias_ms(tp_tm_e tracking_mode);
+float tp_get_alias_ms(tp_tm_e tracking_mode);
 float tp_get_flld_ms(tp_tm_e tracking_mode);
 u8 tp_get_flll_ms(tp_tm_e tracking_mode);
 u8 tp_get_bit_ms(tp_tm_e tracking_mode);
